Fix No pcap_meta info
修復 pcap_meta 的相關資訊並未成功由大端序轉換成小端序，將會導致後續
的程式讀取發生錯誤，因此將讀取到的 buffer 資訊先轉換成 `Flow2img_II_Proxy` 結構
再對該結構內需要轉換成小端序的類型進行轉換。
並且添加該作法的註解。

對於 USTC-II/README.md，刪除 `HAST_Two` 已知錯誤的部分，添加上關於若要傳送給
`HAST_Two` 程式的 Socket 說明。

+++ USTC-II/common.h
@@ -51,7 +51,7 @@ typedef enum {
  */
 typedef struct {
     uint8_t pcap_data[MAX_PKT_LEN * N_PKTS];
-    size_t pcap_size;
+    uint32_t pcap_size;
     char s_ip[INET_ADDRSTRLEN];
     char d_ip[INET_ADDRSTRLEN];
     int s_port;
@@ -65,7 +65,7 @@ typedef struct {
  * 1. fd: socket file descriptor
  * 2. state: 目前的讀取狀態 (讀取 header 或 body)
  * 3. buffer: 用來暫存從 socket 讀進來的資料
- * 4. to_read: 還需要讀取多少 bytes
+ * 4. to_read: 預期需要讀取的 bytes 數量 (需要在下次讀取前需要設定好了)
  * 5. received: 已經接收了多少 bytes
  */
 typedef struct {
